+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed 'delete match' not working [#11](https://github.com/cheshire137/competiwatch-desktop/issues/11)
 - Fixed match edit button not being visible in the dark theme
 - Added links to about page to view changes, view source, and report a bug
+- Show paths on the settings page to the files where Competiwatch stores your data
 
 ## Version 0.1.0